Fixed phone number handler.
 On branch master
 Your branch is ahead of 'origin/master' by 2 commits.
 Changes to be committed:
	modified:   wakebot/processors/__init__.py
	modified:   wakebot/processors/reserve.py

+++ wakebot/processors/reserve.py
@@ -118,7 +118,8 @@ class ReserveProcessor(StatedProcessor):
         text, reply_markup, state, answer = self.create_book_message()
         answer = await message.answer(text, reply_markup=reply_markup,
                                       parse_mode=self.parse_mode)
-        await message.reply_to_message.delete()
+        if message.reply_to_message:
+            await message.reply_to_message.delete()
         await message.delete()
 
         await message.answer(text=reply_text,
